+++ E_Commerce.MvcWebUI/Controllers/ProductController.cs
@@ -20,18 +20,40 @@ namespace E_Commerce.MvcWebUI.Controllers
         }
         public async Task<IActionResult> Index()
         {
-            FirebaseResponse response = await dbcontext.client.GetAsync("Product");
-            dynamic data = JsonConvert.DeserializeObject<dynamic>(response.Body);
-            var list = new List<Product>();
-            foreach (var item in data)
+            var token = HttpContext.Session.GetString("_UserToken");
+            var userid = HttpContext.Session.GetString("_UserId");
+            var userRole = HttpContext.Session.GetString("_UserRole");
+
+            if (token != null && userRole == "admin")
             {
-                list.Add(JsonConvert.DeserializeObject<Product>(((JProperty)item).Value.ToString()));
+                FirebaseResponse response = await dbcontext.client.GetAsync("Product");
+                dynamic data = JsonConvert.DeserializeObject<dynamic>(response.Body);
+                var list = new List<Product>();
+                foreach (var item in data)
+                {
+                    list.Add(JsonConvert.DeserializeObject<Product>(((JProperty)item).Value.ToString()));
+                }
+                return View(list);
             }
-            return View(list);
+            else
+            {
+                return RedirectToAction("SignIn", "Account");
+            }
         }
         public IActionResult Add()
         {
-            return View();
+            var token = HttpContext.Session.GetString("_UserToken");
+            var userid = HttpContext.Session.GetString("_UserId");
+            var userRole = HttpContext.Session.GetString("_UserRole");
+
+            if (token != null && userRole == "admin")
+            {
+                return View();
+            }
+            else
+            {
+                return RedirectToAction("SignIn", "Account");
+            }
         }
         [HttpPost]
         public async Task<IActionResult> Add(Product product, IFormFile file)
@@ -42,12 +64,23 @@ namespace E_Commerce.MvcWebUI.Controllers
             PushResponse responses = await dbcontext.client.PushAsync("Product/", product);
             product.Id = responses.Result.name;
             SetResponse setResponse = await dbcontext.client.SetAsync("Product/" + product.Id, product);
-            return View();
+            return RedirectToAction("Index", "Product");
         }
 
         public IActionResult Edit(string id)
         {
-            return View();
+            var token = HttpContext.Session.GetString("_UserToken");
+            var userid = HttpContext.Session.GetString("_UserId");
+            var userRole = HttpContext.Session.GetString("_UserRole");
+
+            if (token != null && userRole == "admin")
+            {
+                return View();
+            }
+            else
+            {
+                return RedirectToAction("SignIn", "Account");
+            }
         }
 
         [HttpPost]
@@ -64,11 +97,20 @@ namespace E_Commerce.MvcWebUI.Controllers
 
         public async Task<IActionResult> Delete(string id)
         {
-            FirebaseResponse response = await dbcontext.client.DeleteAsync("Product/" + id);
-            Console.WriteLine(response.StatusCode);
-            return View();
-        }
+            var token = HttpContext.Session.GetString("_UserToken");
+            var userid = HttpContext.Session.GetString("_UserId");
+            var userRole = HttpContext.Session.GetString("_UserRole");
 
-     
+            if (token != null && userRole == "admin")
+            {
+                FirebaseResponse response = await dbcontext.client.DeleteAsync("Product/" + id);
+                Console.WriteLine(response.StatusCode);
+                return RedirectToAction("Index", "Product");
+            }
+            else
+            {
+                return RedirectToAction("SignIn", "Account");
+            } 
+        }  
     }
 }
